+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Refactored onboarding service [#4826](https://github.com/rokwire/illinois-app/issues/4826).
 
 ## [6.2.0] - 2025-02-25
 ### Fixed